move waas logic into ExportWaaSModal
